Fix SupplyDemandBarChart to accept List<Double> instead of List<Long>
The AnalysisScreen passes Double values for foreign/institution net buying,
so updated both the chart and marker view to accept Double types.

+++ StockApp/app/src/main/java/com/stockapp/core/ui/component/chart/TechnicalCharts.kt
@@ -1103,8 +1103,8 @@ fun MarketCapOscillatorChart(
 @Composable
 fun SupplyDemandBarChart(
     dates: List<String>,
-    foreignValues: List<Long>,
-    institutionValues: List<Long>,
+    foreignValues: List<Double>,
+    institutionValues: List<Double>,
     modifier: Modifier = Modifier
 ) {
     val context = LocalContext.current
